+++ score.hpp
@@ -28,18 +28,18 @@ namespace Objectives {
 
   class ParametricContext {
   protected:
-    std::vector<float> a_;
-    std::vector<float> b_;
+    std::vector<double> a_;
+    std::vector<double> b_;
     int n_;
-    std::vector<std::vector<float> > a_sums_;
-    std::vector<std::vector<float> > b_sums_;
+    std::vector<std::vector<double> > a_sums_;
+    std::vector<std::vector<double> > b_sums_;
     bool risk_partitioning_objective_;
     bool use_rational_optimization_;
     std::string name_;
 
   public:
-    ParametricContext(std::vector<float> a, 
-		      std::vector<float> b, 
+    ParametricContext(std::vector<double> a, 
+		      std::vector<double> b, 
 		      int n, 
 		      bool risk_partitioning_objective,
 		      bool use_rational_optimization,
@@ -58,19 +58,19 @@ namespace Objectives {
 
     virtual void compute_partial_sums() {};
 
-    virtual float compute_score_multclust(int, int) = 0;
-    virtual float compute_score_multclust_optimized(int, int) = 0;
-    virtual float compute_score_riskpart(int, int) = 0;
-    virtual float compute_score_riskpart_optimized(int, int) = 0;
+    virtual double compute_score_multclust(int, int) = 0;
+    virtual double compute_score_multclust_optimized(int, int) = 0;
+    virtual double compute_score_riskpart(int, int) = 0;
+    virtual double compute_score_riskpart_optimized(int, int) = 0;
 
-    virtual float compute_ambient_score_multclust(float, float) = 0;
-    virtual float compute_ambient_score_riskpart(float, float) = 0;
+    virtual double compute_ambient_score_multclust(double, double) = 0;
+    virtual double compute_ambient_score_riskpart(double, double) = 0;
 
     std::string getName() const { return name_; }
     bool getRiskPartitioningObjective() const { return risk_partitioning_objective_; }
     bool getUseRationalOptimization() const { return use_rational_optimization_; }
 
-    float compute_score(int i, int j) {
+    double compute_score(int i, int j) {
       if (risk_partitioning_objective_) {
 	if (use_rational_optimization_) {
 	  return compute_score_riskpart_optimized(i, j);
@@ -89,7 +89,7 @@ namespace Objectives {
       }
     }
     
-    float compute_ambient_score(float a, float b) {
+    double compute_ambient_score(double a, double b) {
       if (risk_partitioning_objective_) {
 	return compute_ambient_score_riskpart(a, b);
       }
@@ -102,8 +102,8 @@ namespace Objectives {
   class PoissonContext : public ParametricContext {
 
   public:
-    PoissonContext(std::vector<float> a, 
-		   std::vector<float> b, 
+    PoissonContext(std::vector<double> a, 
+		   std::vector<double> b, 
 		   int n, 
 		   bool risk_partitioning_objective,
 		   bool use_rational_optimization) : ParametricContext(a,
@@ -117,30 +117,30 @@ namespace Objectives {
       }
     }
   
-    float compute_score_multclust(int i, int j) override {    
-      float C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
-      float B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
+    double compute_score_multclust(int i, int j) override {    
+      double C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
+      double B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
       return (C>B)? C*std::log(C/B) + B - C : 0.;
     }
 
-    float compute_score_riskpart(int i, int j) override {
-      float C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
-      float B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
+    double compute_score_riskpart(int i, int j) override {
+      double C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
+      double B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
       return C*std::log(C/B);
     }
     
-    float compute_ambient_score_multclust(float a, float b) override {
+    double compute_ambient_score_multclust(double a, double b) override {
       return (a>b)? a*std::log(a/b) + b - a : 0.;
     }
 
-    float compute_ambient_score_riskpart(float a, float b) override {
+    double compute_ambient_score_riskpart(double a, double b) override {
       return a*std::log(a/b);
     }  
 
     void compute_partial_sums() override {
-      float a_cum;
-      a_sums_ = std::vector<std::vector<float> >(n_, std::vector<float>(n_+1, std::numeric_limits<float>::lowest()));
-      b_sums_ = std::vector<std::vector<float> >(n_, std::vector<float>(n_+1, std::numeric_limits<float>::lowest()));
+      double a_cum;
+      a_sums_ = std::vector<std::vector<double> >(n_, std::vector<double>(n_+1, std::numeric_limits<double>::lowest()));
+      b_sums_ = std::vector<std::vector<double> >(n_, std::vector<double>(n_+1, std::numeric_limits<double>::lowest()));
     
       for (int i=0; i<n_; ++i) {
 	a_sums_[i][i] = 0.;
@@ -157,13 +157,13 @@ namespace Objectives {
       }  
     }
 
-    float compute_score_riskpart_optimized(int i, int j) override {
-      float score = a_sums_[i][j]*std::log(a_sums_[i][j]/b_sums_[i][j]);
+    double compute_score_riskpart_optimized(int i, int j) override {
+      double score = a_sums_[i][j]*std::log(a_sums_[i][j]/b_sums_[i][j]);
       return score;
     }
     
-    float compute_score_multclust_optimized(int i, int j) override {
-      float score = (a_sums_[i][j] > b_sums_[i][j]) ? a_sums_[i][j]*std::log(a_sums_[i][j]/b_sums_[i][j]) + b_sums_[i][j] - a_sums_[i][j]: 0.;
+    double compute_score_multclust_optimized(int i, int j) override {
+      double score = (a_sums_[i][j] > b_sums_[i][j]) ? a_sums_[i][j]*std::log(a_sums_[i][j]/b_sums_[i][j]) + b_sums_[i][j] - a_sums_[i][j]: 0.;
       return score;
     }
     
@@ -172,8 +172,8 @@ namespace Objectives {
   class GaussianContext : public ParametricContext {
 
   public:
-    GaussianContext(std::vector<float> a, 
-		    std::vector<float> b, 
+    GaussianContext(std::vector<double> a, 
+		    std::vector<double> b, 
 		    int n, 
 		    bool risk_partitioning_objective,
 		    bool use_rational_optimization) : ParametricContext(a,
@@ -187,30 +187,30 @@ namespace Objectives {
       }
     }
   
-    float compute_score_multclust(int i, int j) override {
-      float C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
-      float B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
+    double compute_score_multclust(int i, int j) override {
+      double C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
+      double B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
       return (C>B)? .5*(std::pow(C,2)/B + B) - C : 0.;
     }
   
-    float compute_score_riskpart(int i, int j) override {
-      float C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
-      float B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
+    double compute_score_riskpart(int i, int j) override {
+      double C = std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.);
+      double B = std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
       return C*C/2./B;
     }
 
-    float compute_ambient_score_multclust(float a, float b) override {
+    double compute_ambient_score_multclust(double a, double b) override {
       return (a>b)? .5*(std::pow(a,2)/b + b) - a : 0.;
     }
 
-    float compute_ambient_score_riskpart(float a, float b) override {
+    double compute_ambient_score_riskpart(double a, double b) override {
       return a*a/2./b;
     }
 
     void compute_partial_sums() override {
-      float a_cum;
-      a_sums_ = std::vector<std::vector<float> >(n_, std::vector<float>(n_+1, std::numeric_limits<float>::lowest()));
-      b_sums_ = std::vector<std::vector<float> >(n_, std::vector<float>(n_+1, std::numeric_limits<float>::lowest()));
+      double a_cum;
+      a_sums_ = std::vector<std::vector<double> >(n_, std::vector<double>(n_+1, std::numeric_limits<double>::lowest()));
+      b_sums_ = std::vector<std::vector<double> >(n_, std::vector<double>(n_+1, std::numeric_limits<double>::lowest()));
     
       for (int i=0; i<n_; ++i) {
 	a_sums_[i][i] = 0.;
@@ -227,13 +227,13 @@ namespace Objectives {
       }  
     }
 
-    float compute_score_multclust_optimized(int i, int j) override {
-      float score = (a_sums_[i][j] > b_sums_[i][j]) ? .5*(std::pow(a_sums_[i][j], 2)/b_sums_[i][j] + b_sums_[i][j]) - a_sums_[i][j] : 0.;
+    double compute_score_multclust_optimized(int i, int j) override {
+      double score = (a_sums_[i][j] > b_sums_[i][j]) ? .5*(std::pow(a_sums_[i][j], 2)/b_sums_[i][j] + b_sums_[i][j]) - a_sums_[i][j] : 0.;
       return score;
     }
     
-    float compute_score_riskpart_optimized(int i, int j) override {
-      float score = a_sums_[i][j]*a_sums_[i][j]/2./b_sums_[i][j];
+    double compute_score_riskpart_optimized(int i, int j) override {
+      double score = a_sums_[i][j]*a_sums_[i][j]/2./b_sums_[i][j];
       return score;
     }
 
@@ -246,8 +246,8 @@ namespace Objectives {
     // XGBoost, e.g.
 
   public:
-    RationalScoreContext(std::vector<float> a,
-			 std::vector<float> b,
+    RationalScoreContext(std::vector<double> a,
+			 std::vector<double> b,
 			 int n,
 			 bool risk_partitioning_objective,
 			 bool use_rational_optimization) : ParametricContext(a,
@@ -262,9 +262,9 @@ namespace Objectives {
     }
 
     void compute_partial_sums() override {
-      float a_cum;
-      a_sums_ = std::vector<std::vector<float> >(n_, std::vector<float>(n_+1, std::numeric_limits<float>::lowest()));
-      b_sums_ = std::vector<std::vector<float> >(n_, std::vector<float>(n_+1, std::numeric_limits<float>::lowest()));
+      double a_cum;
+      a_sums_ = std::vector<std::vector<double> >(n_, std::vector<double>(n_+1, std::numeric_limits<double>::lowest()));
+      b_sums_ = std::vector<std::vector<double> >(n_, std::vector<double>(n_+1, std::numeric_limits<double>::lowest()));
     
       for (int i=0; i<n_; ++i) {
 	a_sums_[i][i] = 0.;
@@ -281,30 +281,30 @@ namespace Objectives {
       }  
     }
   
-    float compute_score_multclust(int i, int j) override {
-      float score = std::pow(std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.), 2) /
+    double compute_score_multclust(int i, int j) override {
+      double score = std::pow(std::accumulate(a_.cbegin()+i, a_.cbegin()+j, 0.), 2) /
 	std::accumulate(b_.cbegin()+i, b_.cbegin()+j, 0.);
       return score;
     }
   
-    float compute_score_riskpart(int i, int j) override {
+    double compute_score_riskpart(int i, int j) override {
       return compute_score_multclust(i, j);
     }
 
-    float compute_score_riskpart_optimized(int i, int j) override {
+    double compute_score_riskpart_optimized(int i, int j) override {
       return compute_score_multclust_optimized(i, j);
     }
 
-    float compute_score_multclust_optimized(int i, int j) override {
-      float score = a_sums_[i][j] / b_sums_[i][j];
+    double compute_score_multclust_optimized(int i, int j) override {
+      double score = a_sums_[i][j] / b_sums_[i][j];
       return score;
     }
 
-    float compute_ambient_score_multclust(float a, float b) override {
+    double compute_ambient_score_multclust(double a, double b) override {
       return a*a/b;
     }
 
-    float compute_ambient_score_riskpart(float a, float b) override {
+    double compute_ambient_score_riskpart(double a, double b) override {
       return a*a/b;
     }
